requirements.txt
absl-py==0.3.0
astor==0.7.1
certifi==2018.8.13
cycler==0.10.0
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
Keras==2.1.3
Markdown==2.6.11
matplotlib==2.1.2
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.14.5
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
tornado==4.5.3
Werkzeug==0.14.1
wincertstore==0.2
